Updating #1892, attempt 1 This is the first attempt to solve #1892.
Here is what's done here:
1. First of all, "restartByReachability" logic was restored. Now the tunnel
reloads properly when we detect network changes. If it's disabled (default)
we restart the DNS proxy and update the tunnel settings (the tunnel settings
are first set to empty so that we could get the proper DNS addresses for the
fallback). If it's enabled, we just cancel the tunnel right away on any
network change. The tunnel is then re-established due to on-demand rules.
This way we can guarantee that the tunnel is fully reinitialized.
2. Second, I've added more logging to NetworkUtils so that we could observe
what are the actual network changes we try to react to.