Improve memory usage in ActivityStatistics.compressTable, fix #1935
The issue was caused by "insertMany", this method forms a large SQL
statement that was simply using too much memory on the tunnel init.
On the other hand, calling "compressTable" in the tunnel process is
important since people may not run the app at all and we'll end up
with a huge uncompressed database.